requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-bootstrap==0.2.4
django-bootstrap4==0.0.6
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-filter==2.0.0
django-rest-swagger==2.2.0
django-summernote==0.8.8.7
django-tables2==1.21.2
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jsonapi==2.5.0
djangorestframework-jwt==1.11.0
djoser==1.2.0
fpdf==1.7.2
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
